Implement Exposed models and Kotlin data classes for tracked players and daily snapshots; modified tracked players model in /db directory.
