+++ backend/src/main/kotlin/plugins/Monitoring.kt
@@ -1,0 +1,13 @@
+package com.brawlpulse.api.plugins
+
+import io.ktor.server.application.*
+import io.ktor.server.plugins.calllogging.*
+import io.ktor.server.request.*
+import org.slf4j.event.*
+
+fun Application.configureMonitoring() {
+    install(CallLogging) {
+        level = Level.INFO
+        filter { call -> call.request.path().startsWith("/") }
+    }
+}
