+++ backend/src/main/kotlin/plugins/Serialization.kt
@@ -1,0 +1,13 @@
+package com.brawlpulse.api.plugins
+
+import io.ktor.server.application.*
+import io.ktor.server.response.*
+import io.ktor.server.routing.*
+
+fun Application.configureSerialization() {
+    routing {
+        get("/json/kotlinx-serialization") {
+            call.respond(mapOf("hello" to "world"))
+        }
+    }
+}
